chore: empire base renames and docs (#15750)
Lot of comments on EmpireBase, GovernanceProposer, and GSEPayload, and
introduces some renames.

Generally, on the EmpireBase now, the IEmperor is used to determine the
current "signaler", who may "signal" on a "payload"; I want to preserve
"proposal" and "vote" for governance, i.e. when a payload is actually
`propose`ed to Governance.sol, it becomes part of a "proposal".

Further, the `_execute` virtual function is renamed to
`_handleRoundWinner`, so you don't "executeProposal" on the EmpireBase
anymore, instead you "submitRoundWinner" (which calls
`_handleRoundWinner`).

Also, in round accounting, `lastVote` is now `lastSignalSlot`, and
`leader` is `payloadWithMostSignals`.